refactor: move schema inspection logic into seperate file
